test(telemetry): OTLP export integration test against a mock collector
Add tests/otel_export.rs: an in-process mock OTLP collector (protobuf
POST /v1/traces and /v1/metrics, the surface a sidecar exposes on
:4318) plus a mock upstream, driving a live router with a [telemetry]
table through the real config -> telemetry::init -> AppState order. One
test function on purpose - telemetry installs process globals
(subscriber, meter provider, propagator) that parallel tests must not
race on. Asserts over the wire:

- the exported span tree (chat_completions -> classify,
  upstream_request) carries the routing decision as typed attributes:
  model selection, complexity, engine, exact prompt/window sizes,
  estimated tokens, upstream status, and upstream-reported token usage
- every span joins the caller's inbound W3C traceparent trace id, and
  the upstream receives a traceparent carrying that trace id with the
  router's span as parent
- no exported attribute contains prompt text (privacy invariant)
- all documented hyper_mcp_router.* metric names arrive
- providers flush on shutdown (the test depends on it)

The test immediately caught a real bug: span fields recorded as u64
export as debug-formatted STRINGS (tracing-opentelemetry overrides
record_i64 but not record_u64), making sizes, token counts, and status
unfilterable as numbers in trace backends. Record them as i64 in proxy
and usage instead.

Dev-deps: opentelemetry-proto (gen-tonic-messages) + prost for payload
decoding - both already in the tree transitively.

+++ src/usage.rs
@@ -66,15 +66,17 @@ pub(crate) fn report_upstream_usage(
             "upstream token usage"
         );
     }
+    // i64: `tracing-opentelemetry` maps i64 to a real OTel integer
+    // attribute, while u64 would export as a debug-formatted STRING.
     let span = tracing::Span::current();
     if let Some(prompt) = usage.prompt_tokens {
-        span.record("prompt_tokens", prompt);
+        span.record("prompt_tokens", prompt as i64);
     }
     if let Some(completion) = usage.completion_tokens {
-        span.record("completion_tokens", completion);
+        span.record("completion_tokens", completion as i64);
     }
     if let Some(total) = usage.total_tokens {
-        span.record("total_tokens", total);
+        span.record("total_tokens", total as i64);
     }
     if let Some(metrics) = usage_metrics {
         let attrs = [attr("model", model.to_string())];
